// src/domain/model/result-starships.model.ts
import { inherits } from "util";

export interface ResultStarshipsModel {
  nombre?: string;
  modelo?: string;
  fabricante?: string;
  costo_en_credito?: string;
  longitud?: string;
  velocidad_maxima_atmosfera?: string;
  tripulacion?: string;
  pasajeros?: string;
  capacidad_carga?: string;
  consumables?: string;
  calificacion_hiperimpulsor?: string;
  MGLT?: string;
  calse_nave_estelar?: string;
  pilotos?: string[];
  peliculas?: string[];
  creado?: string;
  editado?: string;
  url?: string;
}